+++ fixtures/gists-app/app/scripts/client.js
@@ -1,0 +1,1 @@
+export const message = "I'm on the client";