+++ fixtures/gists-app/app/scripts/server.js
@@ -1,0 +1,1 @@
+export const message = "I'm on the server";